feat(reliability): 3c 分片聚合与部分成功(成功策略 + PARTIAL_SUCCESS 状态)
父 execution 与 DAG 节点的终态不再裸判 anyFailed,改由分片终态计数叠加任务
成功策略推导(与旧全成/全败语义完全兼容):

- core:ExecutionStatus 新增 PARTIAL_SUCCESS(isTerminal 含它,ExecutionTransitions
  穷举 switch 同步);新增 SuccessPolicy——NONE/RATIO_PERCENT/MIN_SUCCESS/MAX_FAILURES
  四种阈值的 satisfied + aggregateTerminal(anyCancelled&&failed==0→CANCELED,
  不达标→FAILED,failed==0→SUCCESS,否则 PARTIAL_SUCCESS)。Task 记录追加
  successPolicyType/successPolicyValue 两字段。
- persistence:V26 迁移加 app_task.success_policy_type/value 两列;JdbcTaskRepository
  MAP/create/update 落库回读;parentsNeedingAggregation 返回任务 id 以便取策略。
- server:Reconciler 汇聚父终态(保留 FAILED/CANCELED/SUCCESS 旧 detail 文案,
  PARTIAL_SUCCESS 新文案);DagEngine 节点终态同源推导,PARTIAL_SUCCESS 计为节点
  成功放行 all_success 下游且不触发重试;NotificationFirer 新增
  execution.partial_completed 通知(幂等键 parent:execution.partial_completed:{id});
  ExecutionController RERUNNABLE 含 PARTIAL_SUCCESS;TaskController 白名单/越界 400。
- worker:ExecutorWorkerTest 补 NoopNotificationRepo 接 6 参 Reconciler。
- web:Task/CreateTaskRequest 加策略字段;任务表单加「成功策略」下拉+阈值输入
  (切类型重置阈值防越界 400);StatusBadge 新增 PARTIAL_SUCCESS 琥珀色。

测试:ReconcilerTest 18、DagEngineTest 17(PARTIAL_SUCCESS→下游放行)、
NotificationFirerTest 7、ApiIntegrationTest 109 + worker 全绿;前端 tsc+vite 构建过。

+++ scheduler-persistence/src/main/java/dev/scheduler/persistence/JdbcAuditRepository.java
@@ -68,7 +68,7 @@ public class JdbcAuditRepository implements AuditRepository {
     // (archived_at = 当前最大),避免并发归档误删彼此批次;随后 scheduler_audit_rechain() 重算剩余行,
     // 使删除不破坏取证链(integrity() 仍 verified)。LIMIT 按 id 升序截断单次删除行数,供预算控制。
     return tx.execute(status -> {
-      int copied = jdbc.update("""
+      long copied = jdbc.update("""
           INSERT INTO app_audit_archive (id, operator, action, target_type, target_id, meta, diff, before_meta,
                                          source, occurred_at, prev_hash, chunk_hash, archived_at)
           SELECT id, operator, action, target_type, target_id, meta, diff, before_meta,
